requirements.txt
langchain-community
flask
python-dotenv
langchain_google_genai
langchain_openai
webbrowser
